+++ lib/mixins/sensor.dart
@@ -9,7 +9,6 @@ final Color sensorColor = const Color(0xFFF44336).withOpacity(0.5);
 mixin Sensor<T extends GameComponent> on GameComponent {
   static const _sensorIntervalKey = 'SensorContact';
   int _intervalCallback = 100;
-  GameComponent? componentIncontact;
   bool sensorEnabled = true;
 
   void onContact(T component) {}
@@ -17,21 +16,6 @@ mixin Sensor<T extends GameComponent> on GameComponent {
 
   void setSensorInterval(int intervalCallback) {
     _intervalCallback = intervalCallback;
-  }
-
-  @override
-  void update(double dt) {
-    super.update(dt);
-    if (componentIncontact != null && sensorEnabled) {
-      if (checkInterval(
-        _sensorIntervalKey,
-        _intervalCallback,
-        dt,
-        firstCheckIsTrue: true,
-      )) {
-        onContact(componentIncontact! as T);
-      }
-    }
   }
 
   @override
@@ -46,17 +30,24 @@ mixin Sensor<T extends GameComponent> on GameComponent {
   @override
   void onCollision(Set<Vector2> intersectionPoints, PositionComponent other) {
     if (other is T) {
-      componentIncontact = other;
+      if (sensorEnabled) {
+        print(other.runtimeType);
+        if (checkInterval(
+          _sensorIntervalKey,
+          _intervalCallback,
+          lastDt,
+        )) {
+          onContact(other);
+        }
+      }
     }
     super.onCollision(intersectionPoints, other);
   }
 
   @override
   void onCollisionEnd(PositionComponent other) {
-    if (componentIncontact == other) {
-      componentIncontact = null;
-      onContactExit(other as T);
-      resetInterval(_sensorIntervalKey);
+    if (other is T) {
+      onContactExit(other);
     }
     super.onCollisionEnd(other);
   }